# eliminate `Lazy<string>` from `Address`