This PR:

1. Improves `ActorPath.Parse` performance
2. Reduces `Address` memory allocation