@Arkatufus 2021-06-03
This optimization might actually hurt memory footprint in use cases where there are a lot of actors and address ToString() are never called.