@Arkatufus 2021-06-03
Oh, I didn't notice the changes on the ToString() method, yes, this is good.